Work around the fact that we're faking things
Currently, we're using files (for testing purposes) that we claim start
at a different point in time than they actually do. This works, but now
that detect_files is ran every so often, the endtimes get set to some
time before the starttimes, which means our tests break.

Fix it up by, when updating, using the starttime as found in the
database rather than as just parsed out of the filename, so that we
don't have any further problems anymore